requirements.txt
matplotlib~=3.7.1
numpy~=1.23.5
torch~=2.0.1
requests~=2.28.2
Pillow~=8.4.0
torchvision~=0.15.2